arweb/requirements.txt
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
mysqlclient==1.3.12
pytz==2017.3
